modify scene struct, refactor scenes, refactor game.Update()
this modifies the `Scene` interface to add new methods

- `ID() SceneId`
- `Next() SceneId`
- `SetFreezeFrame(*ebiten.Image)`

the main change to the `Update()` method is to return an error instead of a
`SceneID`.  This change aligns with the ebitengine `game.Update()` method.  As a
result, a new `scene{}` type is addded, which has common fields between all
scenes and new fields for the new methods.

The behavior of the `Scene` interface's `Update()` method is to now set `s.next`
and then `return nil` to set the next scene.  This pattern extends into the
`game.Update()` method, which has a lowered cognitive-complexity rating.

The main pattern in the `game.Update()` method is to run `s.Update()` by
default, and only if the next scene is different than the current:

1. check for pause
2. exit the current scene
3. set the current scene to the next
4. enter the (now current) next scene
5. update again, as usual

A few other adjustments were made, such as:

- using the same variable name for all scenes
- move the nil variable for detecting interface compliance to the top of the the
  files
- other small tweaks

the behavior of the game is exactly the same after this refactor and it lays the
groundwork for changing the game signifigantly as I begin converting this RPG
into a pseudo3d racing game.

Signed-off-by: Jacob Salmela <me@jacobsalmela.com>